feat: redesign landing page to include both a /my and /public route for different projects
