+++ frontend/src/lib/components/general/Banner.svelte
@@ -1,0 +1,5 @@
+<div class="bg-yellowish rounded-2xl flex items-center w-full">
+	<p class="text-lg m-4">
+		<slot />
+	</p>
+</div>
