+++ frontend/src/routes/(app)/(home)/+page.server.ts
@@ -1,0 +1,11 @@
+import { redirect } from '@sveltejs/kit';
+
+export async function load({ cookies }) {
+	const userCookie = cookies.get('loggedIn');
+
+	if (!userCookie) {
+		throw redirect(303, '/public');
+	} else {
+		throw redirect(303, '/my');
+	}
+}
