+++ frontend/src/routes/(app)/(home)/public/+page.server.ts
@@ -1,0 +1,17 @@
+import { getEndpoint } from '$lib/util/util';
+import { OpenAPI, ZenoService } from '$lib/zenoapi';
+
+export async function load({ depends }) {
+	depends('app:projects');
+	depends('app:reports');
+
+	OpenAPI.BASE = `${getEndpoint()}/api`;
+
+	const publicReports = await ZenoService.getPublicReports();
+	const publicProjects = await ZenoService.getPublicProjects();
+
+	return {
+		publicProjects: publicProjects,
+		publicReports: publicReports
+	};
+}
